Add files via upload
Keep getting low R^2 scores, double check my code to see if theres anything wrong, but if not we might as well go with this for the presentation. Also ran a lot of regressions by sector using groupbys. Data is merged with temmperatures corresponding to the state, added as columns Temp1990, Temp1991, etc.